requirements.txt
fastai==2.8.6
fastcore==1.11.2
fastprogress==1.1.3
pillow
streamlit
torch
torchvision
ipython
